restructuring; basic tests
