+++ tests/pytests/data/input/.gitignore
@@ -1,0 +1,3 @@
+*
+!.gitignore
+!README.md
