+++ tests/pytests/data/input/README.md
@@ -1,0 +1,1 @@
+Test data downloaded from zenodo goes in this folder.
