+++ tests/pytests/data/output/total.tsv
@@ -1,0 +1,101 @@
+normal	21614076
+ACACACCA	957619
+ACTACTGT	956113
+GGGGCGAC	1031993
+ACCCGAAA	953522
+GGCCCTAG	1033041
+GGCGCGCA	1033743
+ACCGTCCT	955027
+GGATAGCT	1033487
+AGCACGTA	953145
+GGGGGACC	1034588
+GGTCTTGG	1035622
+GTTTTGGC	1038136
+GAATGGAA	1035362
+ACGGAGAA	955108
+GCTGGGGG	1035464
+GCGAACCC	1034271
+AAACCCCG	957751
+GTATCGAT	1034150
+ACGGATTT	956703
+AGTCTCTC	954078
+ATCGGGGG	954968
+GGTGTCGC	1035138
+ATATCGAG	952700
+ATGCAAGC	956242
+AGCCAGCA	956451
+GGTGGATT	1034351
+AGTACCAT	954224
+GGAGACTG	1033339
+AGGGCAAT	956665
+ACAGAATA	955855
+ATATTTAC	956517
+GTAGTGAT	1034883
+GTACTCCC	1035802
+GATCCGCC	1034840
+ATCATAGT	955420
+ACCGGCCC	954656
+GGGCAATA	1034728
+AGAAGCGA	958082
+AATGTACA	955777
+AGACGATA	953727
+AGCGATGG	958039
+ATCGGTAT	954110
+GCGGAGTT	1033472
+ACCGCACA	954183
+AGGAGATA	956402
+AGCGTCGA	956780
+ATAGGGCC	956105
+GGTTAACA	1033195
+ACTATTTG	955162
+ATCGGGAC	954520
+AATCGGGC	954964
+GACGATTA	1034515
+GAGGCCGG	1033828
+AGATTAGT	956151
+GGCTGTAT	1035321
+AGGCCACT	957262
+GGAGGATT	1035972
+GCACCTCC	1035573
+GCAACCGA	1037662
+ACATCGTT	952277
+GGACTATG	1034823
+AAACGCCT	954376
+ATTCCGGT	956273
+GGTAATTC	1037230
+GCCCTCTC	1033039
+GACCACGG	1031569
+ACGTCGGA	957720
+GCCCATTC	1036101
+GTAGGATT	1033946
+ATGCGGTA	957359
+ACTAGACG	956299
+GCCTAACT	1033407
+AATTAGCC	955433
+GATTGATT	1036081
+AGATCGCT	956010
+GCTCTCCA	1034989
+AAGGGGTA	954521
+ACATGAGA	953597
+GATCTCGC	1035522
+GCTCGGTT	1037580
+GGTGGGAA	1034029
+GCCAGTCG	1034770
+ATTAGCTA	955147
+AGTATTAT	954554
+AGGAGTTA	954289
+ACATCACT	958150
+ATCATACT	956359
+GGCCCCCC	1035345
+GGTCAGGT	1035923
+GATGCCAC	1036401
+GGGCATTC	1035342
+ATCCATGG	956450
+GCTGTACC	1036280
+ACACCCAA	957480
+GCTATTTA	1034242
+GGGGGGGG	1035002
+GAGGCCCC	1034277
+GAGCAGTA	1036894
+GAGACCGA	1035678
+GACAACAA	1035912
